python-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
cachetools==5.5.2
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cryptography==46.0.1
dnspython==2.8.0
et_xmlfile==2.0.0
fastapi==0.115.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
motor==3.3.2
multidict==6.6.4
oauthlib==3.3.1
openpyxl==3.1.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.24.0
xlrd==2.0.1
yarl==1.20.1
